+++ web/app/integration-manager/_merlins-connector/checkout/parsers.js
@@ -1,4 +1,4 @@
-import {extractMagentoShippingStepData, getCheckoutEntityID} from '../../../utils/magento-utils'
+import {extractMagentoShippingStepData} from '../../../utils/magento-utils'
 
 
 export const checkoutShippingParser = ($, $html) => {
@@ -49,13 +49,10 @@ export const parseLocations = (shippingStepData) => {
 }
 
 export const parseCheckoutData = ($response) => {
-    const customerEntityID = getCheckoutEntityID($response)
     const magentoFieldData = extractMagentoShippingStepData($response).getIn(['children', 'shipping-address-fieldset', 'children'])
     const initialValues = parseShippingInitialValues(magentoFieldData)
     const locationsData = parseLocations(magentoFieldData)
     return {
-        // entity_id is used for API calls
-        customerEntityID,
         ...locationsData,
         shipping: {initialValues}
     }
