+++ web/app/integration-manager/_merlins-connector/reducer.js
@@ -1,5 +1,6 @@
 import {handleActions} from 'redux-actions'
-import {receiveFormInfo} from './actions'
+import {receiveFormInfo, receiveEntityID} from './actions'
+import {mergePayloadForActions} from '../../utils/reducer-utils'
 
 import Immutable from 'immutable'
 
@@ -7,7 +8,7 @@ const initialState = Immutable.Map()
 
 // Add Merlin's-specific actions here
 const reducer = handleActions({
-    [receiveFormInfo]: (state, {payload}) => state.mergeDeep(payload)
+    ...mergePayloadForActions(receiveFormInfo, receiveEntityID)
 }, initialState)
 
 export default reducer
