+++ web/app/integration-manager/_merlins-connector/selectors.js
@@ -1,0 +1,4 @@
+import {createGetSelector} from 'reselect-immutable-helpers'
+
+export const getIntegrationManager = ({integrationManager}) => integrationManager
+export const getCustomerEntityID = createGetSelector(getIntegrationManager, 'customerEntityID', '')
