+++ web/app/store/checkout/selectors.js
@@ -1,6 +1,5 @@
 import {getCheckout} from '../../store/selectors'
 import {createGetSelector} from 'reselect-immutable-helpers'
 
-export const getCustomerEntityID = createGetSelector(getCheckout, 'customerEntityID')
 
 export const getEmailAddress = createGetSelector(getCheckout, 'emailAddress')
