+++ web/app/utils/magento-utils.js
@@ -33,7 +33,7 @@ export const getCheckoutConfigObject = ($html) => {
     return {}
 }
 
-export const getCheckoutEntityID = ($html) => {
+export const parseCheckoutEntityID = ($html) => {
     const configObject = getCheckoutConfigObject($html)
 
     return configObject && configObject.quoteData ? configObject.quoteData.entity_id : ''
